requirements.txt
# Together.ai API client
together

# Core ML libraries for image processing
transformers==4.49.0
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
pillow>=9.0.0,<11.0.0

# Web framework
flask>=2.3.0,<3.0.0

# HTTP requests and API handling
requests>=2.28.0,<3.0.0
httpx>=0.24.0,<1.0.0

# Configuration management
pyyaml>=6.0,<7.0
python-dotenv>=1.0.0,<2.0.0

# Data processing
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0

# Optional: Text-to-speech
gtts>=2.3.0,<3.0.0

# Optional: Computer vision utilities
opencv-python>=4.7.0,<5.0.0

# Optional: Streamlit for quick UI
streamlit>=1.25.0,<2.0.0

# Optional: AWS deployment
boto3>=1.26.0,<2.0.0

# Development and testing
pytest>=7.0.0,<8.0.0
python-multipart>=0.0.6,<1.0.0

# Logging and monitoring
loguru>=0.7.0,<1.0.0
# For common.py utilities
ensure>=1.0.2
python-box[all]>=7.0.0
joblib>=1.3.0

einops>=0.6.0
timm>=0.9.10

# For Story Generation
requests>=2.28.0
python-dotenv>=1.0.0

-e .